requirements.txt
streamlit>=1.20.0
boto3>=1.26.0
PyPDF2>=3.0.1
langchain>=0.3.0
langchain-community>=0.3.0
langchain-aws>=0.1.0
langgraph>=0.1.0
opensearch-py>=2.3.0
requests>=2.28.1
aws-cdk-lib>=2.70.0
constructs>=10.1.233
sentence-transformers>=2.2.2
python-dotenv>=1.0.0